requirements.txt
gpxpy
pandas
geopy
matplotlib
seaborn
folium
aiogram
python-dotenv
requests
